==> code/1.1_feature_engineering.ipynb <==
import numpy as np
import pandas as pd

import warnings

# Suppress all warnings
warnings.filterwarnings("ignore")


# data directory

data_dir = '../input/optiver-realized-volatility-prediction/'

def wap1(df):
    wap = (df['bid_price1'] * df['ask_size1'] + df['ask_price1'] * df['bid_size1'])/(df['bid_size1'] + df['ask_size1'])
    return wap
def wap2(df):
    wap = (df['bid_price2'] * df['ask_size2'] + df['ask_price2'] * df['bid_size2'])/(df['bid_size2'] + df['ask_size2'])
    return wap

def log_return(list_stock_prices):
    return np.log(list_stock_prices).diff() 

def realized_volatility(series):
    return np.sqrt(np.sum(series**2))

def preprocessor_book(file_path):
    
    df = pd.read_parquet(file_path)
    
    #calculate return, and other book-related feautures
    df['wap1'] = wap1(df)
    df['log_return1'] = df.groupby('time_id')['wap1'].transform(log_return)
    
    df['wap2'] = wap2(df)
    df['log_return2'] = df.groupby('time_id')['wap2'].transform(log_return)
    
    df['wap_balance'] = abs(df['wap1'] - df['wap2'])/((df['wap1'] + df['wap2'])/2)
    
    df['price_spread1'] = (df['ask_price1'] - df['bid_price1']) / ((df['ask_price1'] + df['bid_price1'])/2)
    df['price_spread2'] = (df['ask_price2'] - df['bid_price2']) / ((df['ask_price2'] + df['bid_price2'])/2)
    
    df['bid_spread'] = (df['bid_price1'] - df['bid_price2'])/((df['bid_price1'] + df['bid_price2'])/2)
    
    
    df['ask_spread'] = (df['ask_price1'] - df['ask_price2'])/((df['ask_price1'] + df['ask_price2'])/2)
    
    df['total_volume1'] = df['ask_size1'] + df['bid_size1']
    df['total_volume2'] = df['ask_size2'] + df['bid_size2']
    
    df['volume_imbalance1'] = abs(df['ask_size1'] - df['bid_size1'])
    df['volume_imbalance2'] = abs(df['ask_size2'] - df['bid_size2'])

    
    #dict for aggregate
    create_feature_dict = {
        'log_return1':[realized_volatility],
        'log_return2':[realized_volatility],
        'wap_balance':[np.mean, np.std],
        'price_spread1':[np.mean, np.std],
        'price_spread2':[np.mean, np.std],
        'bid_spread':[np.mean, np.std],
        'ask_spread':[np.mean, np.std],
        'volume_imbalance1':[np.mean, np.std],
        'total_volume1':[np.mean, np.std],
        'volume_imbalance2':[np.mean, np.std],
        'total_volume2':[np.mean, np.std],  
        'wap1':[np.mean, np.std],
        'wap2':[np.mean, np.std],   
            }

    #####groupby / all seconds
    df_feature = pd.DataFrame(df.groupby(['time_id']).agg(create_feature_dict)).reset_index()
    
    df_feature.columns = ['_'.join(col) for col in df_feature.columns] #time_id is changed to time_id_
        
    
    last_seconds = [300, 200, 100]
    
    for second in last_seconds:
        second = 600 - second 
    
        df_feature_sec = pd.DataFrame(df.query(f'seconds_in_bucket >= {second}').groupby(['time_id']).agg(create_feature_dict)).reset_index()

        df_feature_sec.columns = ['_'.join(col) for col in df_feature_sec.columns] #time_id is changed to time_id_
     
        df_feature_sec = df_feature_sec.add_suffix('_' + str(second))

        df_feature = pd.merge(df_feature,df_feature_sec,how='left',left_on='time_id_',right_on=f'time_id__{second}')
        df_feature = df_feature.drop([f'time_id__{second}'],axis=1)
    
    
    
    
    #create row_id
    stock_id = file_path.split('=')[1]
    df_feature['row_id'] = df_feature['time_id_'].apply(lambda x:f'{stock_id}-{x}')
    df_feature = df_feature.drop(columns = ['time_id_'])

    
    return df_feature

%%time
file_path = data_dir + "book_train.parquet/stock_id=100"
preprocessor_book(file_path)

def preprocessor_trade(file_path):
    
    df = pd.read_parquet(file_path)

    df['total_size'] = df['size']*df['order_count']

    df['log_return'] = df.groupby('time_id')['price'].transform(log_return)


    agg_dictionary = {
        'log_return':[realized_volatility],
        'seconds_in_bucket':[pd.Series.nunique],
        'total_size':[np.sum, np.std]
    }

    df_feature = df.groupby('time_id').agg(agg_dictionary)
    
    df_feature = df_feature.reset_index()
    df_feature.columns = ['_'.join(col) for col in df_feature.columns]


    last_seconds = [300, 200, 100]
    
    for second in last_seconds:
        second = 600 - second 
    
        df_feature_sec = pd.DataFrame(df.query(f'seconds_in_bucket >= {second}').groupby(['time_id']).agg(agg_dictionary)).reset_index()

        df_feature_sec.columns = ['_'.join(col) for col in df_feature_sec.columns] #time_id is changed to time_id_
     
        df_feature_sec = df_feature_sec.add_suffix('_' + str(second))

        df_feature = pd.merge(df_feature,df_feature_sec,how='left',left_on='time_id_',right_on=f'time_id__{second}')
        df_feature = df_feature.drop([f'time_id__{second}'],axis=1)

    

    df_feature = df_feature.add_prefix('trade_')
    stock_id = file_path.split('=')[1]
    df_feature['row_id'] = df_feature['trade_time_id_'].apply(lambda x:f'{stock_id}-{x}')
    df_feature = df_feature.drop(['trade_time_id_'],axis=1)
    
    return df_feature

    

    

    

    

%%time
file_path = data_dir + "trade_train.parquet/stock_id=0"
preprocessor_trade(file_path)

def feature_gen(list_stock_ids, is_train = True):

    def feature_gen_stock(stock_id):
        if is_train:
            file_path_book = data_dir + "book_train.parquet/stock_id=" + str(stock_id)
            file_path_trade = data_dir + "trade_train.parquet/stock_id=" + str(stock_id)
        else:
            file_path_book = data_dir + "book_test.parquet/stock_id=" + str(stock_id)
            file_path_trade = data_dir + "trade_test.parquet/stock_id=" + str(stock_id)
       
        df_stock = pd.merge(preprocessor_book(file_path_book),preprocessor_trade(file_path_trade),on='row_id',how='left')

        return df_stock

    df_list = []

    for stock_id in list_stock_ids:
        df_list.append(feature_gen_stock(stock_id))

    return pd.concat(df_list, ignore_index = True)

    
    
    

%%time
list_stock_ids = [0,1]
feature_gen(list_stock_ids, is_train = True)

train = pd.read_csv(data_dir + 'train.csv')

train.head(n=10)

list_of_stocks = train.stock_id.unique()

%%time
df_train = feature_gen(list_stock_ids= list_of_stocks, is_train = True)

output_path = "df_train.csv"
df_train.to_csv(output_path, index=False)

df_train.head(n=10)

df_train.info()

df_train.tail()
