==> code/3_lgbm_model.ipynb <==
# LightGBM Prediction Model

In this notebook, we train a LightGB model for predicting the realized volatility. We will use the features generated in the feature engineering notebook and teh temporal structure we recovered in the recovering_time notebook to train the model.

# libraries and settings
import pandas as pd
import numpy as np
import lightgbm as lgb
import optuna
import warnings
warnings.filterwarnings('ignore')

# reading the training data 
df_train = pd.read_csv('../data/df_train.csv')
train = pd.read_csv('../data/train.csv')
time_order_df = pd.read_csv('../data/time_order.csv')

df_train['target'] = train['target']
df_train.dropna(inplace=True)

df_train['stock_id'] = df_train['row_id'].apply(lambda x: int(x.split('-')[0])).astype('category')
df_train['time_id'] = df_train['row_id'].apply(lambda x: int(x.split('-')[1]))
df_train = pd.merge(df_train, time_order_df, on = 'time_id')
df_train.sort_values('time_id_ordered', inplace = True)

## Training and Fine-Tuning the LGBM model

We use the RMSPE loss function. Note that we use the time order we recovered in the time recovery notebook to split our data. More specifically, we use the first 80% of the time period as the training and the remaining 20% as the validation set. This helps us to avoid future looking and get better generalization on real time data.

We use optuna to optimize the gradient boosted tree's paramters.

best_model_container = {'model': None}

# --------------------------
# RMSPE loss (gradient + hessian)
# --------------------------
def rmspe_objective(y_pred, dataset):
    y_true = dataset.get_label()
    eps = 0
    residual = (y_true - y_pred) / (y_true + eps)

    grad = -2.0 * residual / (y_true + eps)
    hess = 2.0 / (y_true + eps) ** 2

    return grad, hess

# --------------------------
# RMSPE evaluation function
# --------------------------
def rmspe_eval(y_pred, dataset):
    y_true = dataset.get_label()
    eps = 0
    pct_error = (y_true - y_pred) / (y_true + eps)
    score = np.sqrt(np.mean(pct_error ** 2))
    return 'RMSPE', score, False  


def objective(trial):

    params = {
        'objective': rmspe_objective,
        'metric': 'none',  
        'verbosity': -1,
        'learning_rate': trial.suggest_float('learning_rate', 0.005, 0.1, log=True),
        'num_leaves': trial.suggest_int('num_leaves', 16, 128),
        'max_depth': trial.suggest_int('max_depth', 3, 7),
        'min_data_in_leaf': trial.suggest_int('min_data_in_leaf', 10, 100),
        'feature_fraction': trial.suggest_float('feature_fraction', 0.6, 1.0),
        'bagging_fraction': trial.suggest_float('bagging_fraction', 0.5, 0.8),
        'bagging_freq': trial.suggest_int('bagging_freq', 1, 10),
        'lambda_l1': trial.suggest_float('lambda_l1', 0.0, 5.0),
        'lambda_l2': trial.suggest_float('lambda_l2', 0.0, 5.0),
        'num_boost_round': trial.suggest_int('num_boost_round', 600, 1000),
        'early_stopping_rounds': 50
    }

    # dividing the data into training and validation datasets, according to the time order
    df_train.sort_values('time_id_ordered', inplace = True)
    t_train = max(df_train['time_id_ordered'])*0.8
    X_train = df_train[df_train['time_id_ordered'] <= t_train].drop(columns=['target', 'row_id', 'time_id', 'time_id_ordered'])
    X_val = df_train[df_train['time_id_ordered'] > t_train].drop(columns=['target', 'row_id', 'time_id', 'time_id_ordered'])
    y_train = df_train[df_train['time_id_ordered'] <= t_train]['target'] 
    y_val = df_train[df_train['time_id_ordered'] > t_train]['target'] 

    categorical_features = ['stock_id']

    lgb_train = lgb.Dataset(
        X_train, 
        y_train, 
        categorical_feature = categorical_features
    )
    
    lgb_val = lgb.Dataset(
        X_val, 
        y_val, 
        categorical_feature = categorical_features, 
        reference = lgb_train
    )

    model = lgb.train(
        params,
        lgb_train,
        valid_sets=[lgb_val],
        feval= rmspe_eval
    )


    def rmspe(y_true, y_pred, eps = 0):
      pct_error = (y_true - y_pred) / (y_true + eps)
      return np.sqrt(np.mean(pct_error ** 2))


    y_val_pred = model.predict(X_val, num_iteration = model.best_iteration)
    rmspe_score = rmspe(y_val, y_val_pred)  

    trial.set_user_attr("model", model)
        
    return rmspe_score


study = optuna.create_study(direction='minimize')
study.optimize(objective, n_trials=50)

print("Best RMSPE:", study.best_value)
print("Best Parameters:")
print(study.best_params)


best_model = study.best_trial.user_attrs["model"]

best_model.save_model('../models/lgbm_model.txt')
